Arel::Nodes::Factorial
