+++ lib/arel/extensions/nodes.rb
@@ -5,16 +5,6 @@
 module Arel
   module Nodes
     # https://www.postgresql.org/docs/9.4/functions-math.html
-    class Factorial < Arel::Nodes::Unary
-      attr_accessor :prefix
-
-      def initialize(expr, prefix)
-        super(expr)
-        @prefix = prefix
-      end
-    end
-
-    # https://www.postgresql.org/docs/9.4/functions-math.html
     class SquareRoot < Arel::Nodes::UnaryOperation
       def initialize(operand)
         super('|/', operand)
